Added all main functions, ready for design
